% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnmrkt-package.R
\docType{package}
\name{nnmrkt-package}
\alias{nnmrkt}
\alias{nnmrkt-package}
\title{nnmrkt: Neural Networks for marketing analysis}
\description{
Neural Networks for marketing analysis
}
\author{
\strong{Maintainer}: Takashi J Ozaki \email{tjo@marketing.tjo.com}

}
\keyword{internal}
