% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{nnmrktPreprocessRecommendData}
\alias{nnmrktPreprocessRecommendData}
\title{Preprocess a recommendation dataset}
\usage{
nnmrktPreprocessRecommendData()
}
\value{

}
\description{
Preprocess a recommendation dataset
}
